+++ src/devcenter/azext_devcenter/tests/latest/test_devcenter_scenario.py
@@ -1647,6 +1647,150 @@ class DevcenterScenarioTest(ScenarioTest):
         )
 
 
+    def test_project_policy_scenario(self):
+        self.kwargs.update(
+            {
+                "rg" : "cli-test",
+                "devcenterName": "cli-test-devcenter",
+                "location": "centraluseuap",
+                "projectPolicyName": "testProjectPolicy"
+            }
+        )
+
+        self.cmd(
+            "az devcenter admin project-policy list "
+            '--resource-group "{rg}" '
+            '--dev-center "{devcenterName}" ',
+            checks=[
+                self.check("length(@)", 0),
+            ],
+        )
+
+        self.cmd(
+            "az devcenter admin project-policy create "
+            '--dev-center "{devcenterName}" '
+            '--name "{projectPolicyName}" '
+            '--resource-group "{rg}" '
+            '--resource-policies \'[{{"resources": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/cli-test/providers/Microsoft.DevCenter/devcenters/cli-test-devcenter/attachednetworks/amlim-nc"}}]\' '
+            '--scopes "[\"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/cli-test/providers/Microsoft.DevCenter/projects/cli-test-project\"]" ',
+            checks=[
+                self.check("name", "{projectPolicyName}"),
+                self.check("resourceGroup", "{rg}"),
+                self.check("scopes[0]", "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/cli-test/providers/Microsoft.DevCenter/projects/cli-test-project"),
+                self.check("resourcePolicies[0].resources", "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/cli-test/providers/Microsoft.DevCenter/devcenters/cli-test-devcenter/attachednetworks/amlim-nc"),
+            ],
+        )
+
+        self.cmd(
+            "az devcenter admin project-policy list "
+            '--resource-group "{rg}" '
+            '--dev-center "{devcenterName}" ',
+            checks=[
+                self.check("length(@)", 1),
+                self.check("[0].name", "{projectPolicyName}"),
+            ],
+        )
+
+        self.cmd(
+            "az devcenter admin project-policy show "
+            '--dev-center "{devcenterName}" '
+            '--name "{projectPolicyName}" '
+            '--resource-group "{rg}" ',
+            checks=[
+                self.check("name", "{projectPolicyName}"),
+                self.check("resourceGroup", "{rg}"),
+                self.check("scopes[0]", "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/cli-test/providers/Microsoft.DevCenter/projects/cli-test-project"),
+                self.check("resourcePolicies[0].resources", "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/cli-test/providers/Microsoft.DevCenter/devcenters/cli-test-devcenter/attachednetworks/amlim-nc"),
+            ],
+        )
+
+        self.cmd(
+            "az devcenter admin project-policy update "
+            '--dev-center "{devcenterName}" '
+            '--name "{projectPolicyName}" '
+            '--resource-group "{rg}" '
+            '--resource-policies \'[{{"resources": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/cli-test/providers/Microsoft.DevCenter/devcenters/cli-test-devcenter/attachednetworks/amlim-nc-eastus"}}]\' '
+            '--scopes "[\"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/cli-test/providers/Microsoft.DevCenter/projects/cli-test-project2\"]" ',
+            checks=[
+                self.check("name", "{projectPolicyName}"),
+                self.check("resourceGroup", "{rg}"),
+                self.check("scopes[0]", "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/cli-test/providers/Microsoft.DevCenter/projects/cli-test-project2"),
+                self.check("resourcePolicies[0].resources", "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/cli-test/providers/Microsoft.DevCenter/devcenters/cli-test-devcenter/attachednetworks/amlim-nc-eastus"),
+            ],
+        )
+
+        self.cmd(
+            "az devcenter admin project-policy delete --yes "
+            '--dev-center "{devcenterName}" '
+            '--name "{projectPolicyName}" '
+            '--resource-group "{rg}"'
+        )
+
+        self.cmd(
+            "az devcenter admin project-policy list "
+            '--resource-group "{rg}" '
+            '--dev-center "{devcenterName}" ',
+            checks=[
+                self.check("length(@)", 0),
+            ],
+        )
+
+    def test_image_definition_scenario(self):
+        self.kwargs.update(
+            {
+                "rg" : "cli-test",
+                "projectName": "cli-test-project",
+                "catalogName": "image",
+                "imageDefName": "project-sample-2"
+            }
+        )
+
+        self.cmd(
+            "az devcenter admin image-definition list "
+            '--resource-group "{rg}" '
+            '--project "{projectName}" '
+            '--catalog-name "{catalogName}" ',
+            checks=[
+                self.check("length(@)", 1),
+                self.check("[0].name", "{imageDefName}"),
+            ],
+        )
+
+    def test_project_image_scenario(self):
+        self.kwargs.update(
+            {
+                "rg" : "cli-test",
+                "projectName": "cli-test-project"
+            }
+        )
+
+        self.cmd(
+            "az devcenter admin project-image list "
+            '--resource-group "{rg}" '
+            '--project "{projectName}" ',
+            checks=[
+                self.check("length(@)", 28),
+            ],
+        )
+    
+    def test_project_sku_scenario(self):
+        self.kwargs.update(
+            {
+                "rg" : "cli-test",
+                "projectName": "cli-test-project"
+            }
+        )
+
+        self.cmd(
+            "az devcenter admin project-sku list "
+            '--resource-group "{rg}" '
+            '--project "{projectName}" ',
+            checks=[
+                self.check("length(@)", 11),
+            ],
+        )
+
+
 @record_only()
 @try_manual
 class DevcenterDataPlaneScenarioTest(ScenarioTest):