+++ src/network-manager/HISTORY.rst
@@ -2,6 +2,11 @@
 
 Release History
 ===============
+1.3.1
++++++
+* Fix argument regex requirement (fix: https://github.com/Azure/azure-rest-api-specs/pull/31130)
+* Stable version
+
 1.3.0
 +++++
 * Add new cmd `az network manager ipam-pool`